3 useful tricks to update github forks
1. a shallow fork clone that keeps all branches with 1 depth and inits submodules (non shallow)
2. a merge and rebase strategy to update the local and origin forks master with upstream master. This is useful for the workflow where you branch from master to create features, being able to easily update that master to the last unspoiled upstream is important. The difference between the merge and rebase strategies is that one has a merge commit (which is annoying when performing multiple pull requests after a upstream update) and the other rewrites history (and thus is inadvisable if you have someone cloning your origin master). 

The shallow fork in 1. works with the strategies in 2. in projects i tried. I'm not a git guru or even a specialist, but i've found it's easy to screw up on updating from upstream with a shallow fork because of the submodules, and was looking for some guidance. This may help the few people that use navi.